fix: cors callback after execution
