+++ api/src/constants/env.ts
@@ -13,4 +13,5 @@ export const {
   INSTAMOJO_CLIENT_ID,
   INSTAMOJO_CLIENT_SECRET,
   INSTAMOJO_BASE_URL,
+  ALLOWED_CLIENT_URL,
 } = process.env;
